requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
chardet==3.0.4
discord.py==1.6.0
idna==3.1
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
pylint==2.6.2
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.6.3
